add "cloudflare:email" handler with RuntimeHooks
