+++ src/presets/cloudflare/runtime/cloudflare-module.ts
@@ -1,7 +1,7 @@
 import "#internal/nitro/virtual/polyfill";
 import { requestHasBody } from "#internal/nitro/utils";
 import { nitroApp } from "#internal/nitro/app";
-import { runCronTasks, runEmailTask, useRuntimeConfig } from "#internal/nitro";
+import { runCronTasks, useRuntimeConfig } from "#internal/nitro";
 import { getPublicAssetMeta } from "#internal/nitro/virtual/public-assets";
 
 import { withoutBase } from "ufo";
@@ -100,19 +100,8 @@ export default {
     if (import.meta._tasks) {
       (globalThis as any).__env__ = env;
 
-      const domain = (event.to as string).split("@")[1];
-
       context.waitUntil(
-        runEmailTask(domain, {
-          context: {
-            ...event,
-            cloudflare: {
-              env,
-              context,
-            },
-          },
-          payload: {},
-        })
+        nitroApp.hooks.callHook("cloudflare:email", event, { env, context })
       );
     }
   },
