+++ src/runtime/virtual/tasks.ts
@@ -1,5 +1,4 @@
 import type { Task, TaskMeta } from "../task";
-import { EmailMeta, EmailTask } from "#internal/nitro/emailtask";
 
 export const tasks: Record<
   string,
@@ -7,12 +6,3 @@ export const tasks: Record<
 > = {};
 
 export const scheduledTasks: false | { cron: string; tasks: string[] }[] = [];
-
-export const emailTasks: Record<
-  string,
-  { resolve?: () => Promise<EmailTask>; meta: EmailMeta }
-> = {};
-
-export const scheduledEmailTasks:
-  | false
-  | { domain: string; tasks: string[] }[] = [];
